requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.1
aiohttp==3.9.5
groq==0.9.0
httpx==0.25.0
gunicorn==21.2.0